UI: Enable Replay Buffer in Advanced Mode
Replay buffer is currently only supported in simple mode. This change adds
support to advanced mode and is, for the most part, a mirror of the
addition of replay buffer to simple mode.

There are some things that were left out, such as max time, since advanced
mode should always have a target bitrate. As per Jim's request, the
options are in a new tab and the buffer is not available with lossless or
FFmpeg enabled.

Mantis-bug: https://obsproject.com/mantis/view.php?id=792